dom/html/test/test_hash_encoded.html

<!doctype html>
<html>
<head>
<title>Test link.hash attribute</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
</head>
<body>

<a id="target1" href="http://www.example.com/#q=♥"></a>
<a id="target2" href="http://www.example.com/#q=â¥"></a>
<a id="target3" href="http://www.example.com/#/search/%23important"></a>

<pre id="test">

<script>
  // Tests Link::GetHash

  var target = document.getElementById("target1");
  is(target.hash, '#q=%E2%99%A5', 'Unexpected hash');

  target = document.getElementById("target2");
  is(target.hash, '#q=%C3%A2%C2%A5', 'Unexpected hash');

  target = document.getElementById("target3");
  is(target.hash, '#/search/%23important', 'Unexpected hash');
</script>

<script>
  // Tests URL::GetHash

  var url = new URL("http://www.example.com/#q=♥")
  is(url.hash, '#q=%E2%99%A5', 'Unexpected hash');

  url = new URL("http://www.example.com/#q=â¥")
  is(url.hash, '#q=%C3%A2%C2%A5', 'Unexpected hash');

  url = new URL("http://www.example.com/#/search/%23important")
  is(url.hash, '#/search/%23important', 'Unexpected hash');
</script>

<script>
  // Tests nsLocation::GetHash

  window.history.pushState(1, document.title, '#q=♥');
  is(location.hash,'#q=%E2%99%A5', 'Unexpected hash');

  window.history.pushState(1, document.title, '#q=â¥');
  is(location.hash,'#q=%C3%A2%C2%A5', 'Unexpected hash');

  window.history.pushState(1, document.title, '#/search/%23important');
  is(location.hash,'#/search/%23important', 'Unexpected hash');
</script>

</pre>
</body>
</html>
